Split index.tsx into components and interface.ts
