+++ src/components/Square.tsx
@@ -1,0 +1,16 @@
+import { ISquare } from '../interface';
+
+interface SquareProps {
+  value: ISquare;
+  onClick: () => void;
+}
+
+function Square(props: SquareProps) {
+  return (
+    <button className="square" onClick={props.onClick}>
+      {props.value}
+    </button>
+  );
+}
+
+export default Square;
