+++ src/interface.ts
@@ -1,0 +1,4 @@
+export type ISquare = 'X' | 'O' | null;
+export interface History {
+  squares: ISquare[];
+}
